requirements/dev.txt
# SHA1:b7eb14c29369180f5bff0c1128d4b718c90123de
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r test.txt
-e file:./build/dependencies/baseframe
    # via -r requirements/base.in
-e file:./build/dependencies/coaster
    # via
    #   -r requirements/base.in
    #   baseframe
astroid==2.15.6
    # via pylint
bandit==1.7.5
    # via -r requirements/dev.in
black==23.7.0
    # via -r requirements/dev.in
build==0.10.0
    # via pip-tools
cattrs==22.1.0
    # via reformat-gherkin
cfgv==3.3.1
    # via pre-commit
cssbeautifier==1.14.9
    # via djlint
dill==0.3.7
    # via pylint
distlib==0.3.7
    # via virtualenv
djlint==1.32.1
    # via -r requirements/dev.in
editorconfig==0.12.3
    # via
    #   cssbeautifier
    #   jsbeautifier
flake8==6.1.0
    # via
    #   -r requirements/dev.in
    #   flake8-annotations
    #   flake8-assertive
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-comprehensions
    #   flake8-docstrings
    #   flake8-isort
    #   flake8-mutable
    #   flake8-print
    #   pep8-naming
flake8-annotations==3.0.1
    # via -r requirements/dev.in
flake8-assertive==2.1.0
    # via -r requirements/dev.in
flake8-blind-except==0.2.1
    # via -r requirements/dev.in
flake8-bugbear==23.7.10
    # via -r requirements/dev.in
flake8-builtins==2.1.0
    # via -r requirements/dev.in
flake8-comprehensions==3.14.0
    # via -r requirements/dev.in
flake8-docstrings==1.7.0
    # via -r requirements/dev.in
flake8-isort==6.0.0
    # via -r requirements/dev.in
flake8-logging-format==0.9.0
    # via -r requirements/dev.in
flake8-mutable==1.2.0
    # via -r requirements/dev.in
flake8-plugin-utils==1.3.3
    # via flake8-pytest-style
flake8-print==5.0.0
    # via -r requirements/dev.in
flake8-pytest-style==1.7.2
    # via -r requirements/dev.in
flask-debugtoolbar==0.13.1
    # via -r requirements/dev.in
gherkin-official==24.0.0
    # via reformat-gherkin
gitdb==4.0.10
    # via gitpython
gitpython==3.1.32
    # via bandit
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
identify==2.5.26
    # via pre-commit
isort==5.12.0
    # via
    #   -r requirements/dev.in
    #   flake8-isort
    #   pylint
jsbeautifier==1.14.9
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.14
    # via djlint
lazy-object-proxy==1.9.0
    # via astroid
lxml-stubs==0.4.0
    # via -r requirements/dev.in
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy==1.4.1
    # via -r requirements/dev.in
mypy-json-report==1.0.4
    # via -r requirements/dev.in
nodeenv==1.8.0
    # via pre-commit
pathspec==0.11.2
    # via
    #   black
    #   djlint
pbr==5.11.1
    # via stevedore
pep8-naming==0.13.3
    # via -r requirements/dev.in
pip-compile-multi==2.6.3
    # via -r requirements/dev.in
pip-tools==7.1.0
    # via pip-compile-multi
platformdirs==3.10.0
    # via
    #   black
    #   pylint
    #   virtualenv
po2json==0.2.2
    # via -r requirements/dev.in
pre-commit==3.3.3
    # via -r requirements/dev.in
pycodestyle==2.11.0
    # via
    #   flake8
    #   flake8-print
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.1.0
    # via flake8
pylint==2.17.5
    # via -r requirements/dev.in
pyproject-hooks==1.0.0
    # via build
pyupgrade==3.10.1
    # via -r requirements/dev.in
reformat-gherkin==3.0.1
    # via -r requirements/dev.in
ruff==0.0.281
    # via -r requirements/dev.in
smmap==5.0.0
    # via gitdb
snowballstemmer==2.2.0
    # via pydocstyle
stevedore==5.1.0
    # via bandit
tokenize-rt==5.2.0
    # via pyupgrade
toposort==1.10
    # via pip-compile-multi
types-chevron==0.14.2.5
    # via -r requirements/dev.in
types-geoip2==3.0.0
    # via -r requirements/dev.in
types-ipaddress==1.0.8
    # via types-maxminddb
types-maxminddb==1.5.0
    # via types-geoip2
types-mock==5.1.0.1
    # via -r requirements/dev.in
types-pyopenssl==23.2.0.2
    # via types-redis
types-python-dateutil==2.8.19.14
    # via -r requirements/dev.in
types-pytz==2023.3.0.0
    # via -r requirements/dev.in
types-redis==4.6.0.3
    # via -r requirements/dev.in
types-requests==2.31.0.2
    # via -r requirements/dev.in
types-urllib3==1.26.25.14
    # via types-requests
virtualenv==20.24.2
    # via pre-commit
wcwidth==0.2.6
    # via reformat-gherkin
wheel==0.41.0
    # via pip-tools
wrapt==1.15.0
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
